unified_document_service/requirements.txt
fastapi
uvicorn
pydantic
python-docx
lxml
openpyxl
regex
python-multipart
